LayoutTests/inspector/audit/basic-string-expected.txt

Testing the basic functionality of audits.


== Running test suite: Audit.Basic
-- Running test case: Audit.Basic.String.Pass
Testing value `"pass"`...
PASS: Result should be "pass".

-- Running test case: Audit.Basic.String.Warn
Testing value `"warn"`...
PASS: Result should be "warn".

-- Running test case: Audit.Basic.String.Fail
Testing value `"fail"`...
PASS: Result should be "fail".

-- Running test case: Audit.Basic.String.Error
Testing value `"error"`...
PASS: Result should be "error".

-- Running test case: Audit.Basic.String.Unsupported
Testing value `"unsupported"`...
PASS: Result should be "unsupported".
